requirements.txt
# CPU-only TF (smaller); if you need GPU, install the proper build instead
tensorflow-cpu==2.15.0
tensorboard==2.15.2

# libs compatible with Py 3.11 and TF 2.15
numpy>=1.26,<2.0
pandas>=2.0,<2.3
scikit-learn>=1.4,<1.6
matplotlib>=3.8,<3.9


streamlit>=1.32,<1.39


# (optional) avoid protobuf conflicts between TF and Streamlit
protobuf>=3.20.3,<4

